refactor: share createIsBuiltin
